swtpm_cert: Cast parameter passed to isspace to int (Cygwin)
Cygwin requires an explicit cast of the char parameter passed to isspace
to int to fix this issue:

ek-cert.c: In function ‘main’:
ek-cert.c:1605:28: error: array subscript has type ‘char’ [-Werror=char-subscripts]
 1605 |             while (isspace(*token))
      |                            ^~~~~~

+++ src/swtpm_cert/ek-cert.c
@@ -1602,7 +1602,7 @@ main(int argc, char *argv[])
 
         token = strtok(s, ",");
         do {
-            while (isspace(*token))
+            while (isspace((int)*token))
                 token++;
             equal = strchr(token, '=');
             if (equal) {
